Consolidates auth code in features
