+++ apps/web/pages/api/book/event.ts
@@ -1,12 +1,11 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import handleNewBooking from "@calcom/features/bookings/lib/handleNewBooking";
-import { getServerSession } from "@calcom/lib/auth";
 import { defaultResponder } from "@calcom/lib/server";
 
 async function handler(req: NextApiRequest & { userId?: number }, res: NextApiResponse) {
-  const session = await getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+  const session = await getServerSession({ req, res });
   /* To mimic API behavior and comply with types */
   req.userId = session?.user?.id || -1;
   const booking = await handleNewBooking(req, {
