+++ apps/web/pages/api/trpc/[trpc].ts
@@ -1,10 +1,9 @@
 /**
  * This file contains tRPC's HTTP response handler
  */
-import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 import { z } from "zod";
 
-import { getServerSession } from "@calcom/lib/auth";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import * as trpcNext from "@calcom/trpc/server/adapters/next";
 import { createContext as createTrpcContext } from "@calcom/trpc/server/createContext";
 import { appRouter } from "@calcom/trpc/server/routers/_app";
@@ -15,7 +14,7 @@ export default trpcNext.createNextApiHandler({
    * @link https://trpc.io/docs/context
    */
   createContext: ({ req, res }) => {
-    const sessionGetter = () => getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+    const sessionGetter = () => getServerSession({ req, res });
 
     return createTrpcContext({ req, res }, sessionGetter);
   },
