+++ apps/web/pages/auth/forgot-password/index.tsx
@@ -4,11 +4,10 @@ import { getCsrfToken } from "next-auth/react";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 import type { SyntheticEvent } from "react";
 import React from "react";
 
-import { getServerSession } from "@calcom/lib/auth";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { Button, EmailField } from "@calcom/ui";
 
@@ -136,7 +135,7 @@ export default function ForgotPassword({ csrfToken }: { csrfToken: string }) {
 export const getServerSideProps = async (context: GetServerSidePropsContext) => {
   const { req, res } = context;
 
-  const session = await getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+  const session = await getServerSession({ req, res });
 
   if (session) {
     res.writeHead(302, { Location: "/" });
