+++ apps/web/server/lib/ssr.ts
@@ -1,8 +1,7 @@
 import type { GetServerSidePropsContext } from "next";
-import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 import superjson from "superjson";
 
-import { getServerSession } from "@calcom/lib/auth";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { createProxySSGHelpers } from "@calcom/trpc/react/ssg";
 import { createContext } from "@calcom/trpc/server/createContext";
 import { appRouter } from "@calcom/trpc/server/routers/_app";
@@ -16,7 +15,7 @@ import { appRouter } from "@calcom/trpc/server/routers/_app";
 export async function ssrInit(context: GetServerSidePropsContext) {
   const { req, res } = context;
 
-  const sessionGetter = () => getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+  const sessionGetter = () => getServerSession({ req, res });
 
   const ctx = await createContext(context, sessionGetter);
