+++ packages/features/auth/lib/hashPassword.ts
@@ -1,0 +1,6 @@
+import { hash } from "bcryptjs";
+
+export async function hashPassword(password: string) {
+  const hashedPassword = await hash(password, 12);
+  return hashedPassword;
+}
