+++ packages/features/auth/lib/identityProviderNameMap.ts
@@ -1,0 +1,7 @@
+import { IdentityProvider } from "@prisma/client";
+
+export const identityProviderNameMap: { [key in IdentityProvider]: string } = {
+  [IdentityProvider.CAL]: "Cal",
+  [IdentityProvider.GOOGLE]: "Google",
+  [IdentityProvider.SAML]: "SAML",
+};
