+++ packages/features/auth/lib/verifyPassword.ts
@@ -1,0 +1,6 @@
+import { compare } from "bcryptjs";
+
+export async function verifyPassword(password: string, hashedPassword: string) {
+  const isValid = await compare(password, hashedPassword);
+  return isValid;
+}
