+++ packages/features/ee/teams/api/upgrade.ts
@@ -3,8 +3,8 @@ import type Stripe from "stripe";
 import { z } from "zod";
 
 import { getRequestedSlugError } from "@calcom/app-store/stripepayment/lib/team-billing";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import stripe from "@calcom/features/ee/payments/server/stripe";
-import { getSession } from "@calcom/lib/auth";
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { HttpError } from "@calcom/lib/http-error";
 import { defaultHandler, defaultResponder } from "@calcom/lib/server";
@@ -64,7 +64,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     closeComUpdateTeam(prevTeam, team);
   }
 
-  const session = await getSession({ req });
+  const session = await getServerSession({ req, res });
 
   if (!session) return { message: "Team upgraded successfully" };
