+++ packages/prisma/seed.ts
@@ -1,11 +1,12 @@
-import { BookingStatus, MembershipRole, Prisma, UserPermissionRole } from "@prisma/client";
+import type { Prisma, UserPermissionRole } from "@prisma/client";
+import { BookingStatus, MembershipRole } from "@prisma/client";
 import { uuid } from "short-uuid";
 
 import dailyMeta from "@calcom/app-store/dailyvideo/_metadata";
 import googleMeetMeta from "@calcom/app-store/googlevideo/_metadata";
 import zoomMeta from "@calcom/app-store/zoomvideo/_metadata";
 import dayjs from "@calcom/dayjs";
-import { hashPassword } from "@calcom/lib/auth";
+import { hashPassword } from "@calcom/features/auth/lib/hashPassword";
 import { DEFAULT_SCHEDULE, getAvailabilityFromSchedule } from "@calcom/lib/availability";
 
 import prisma from ".";
